Filled AnnouncementDto with necessary methods
